move integration test for clients into sep. module
